fix: jump
Пробрасываем напрямую инпут, поправил комментарии
