--- Assets/Scripts/Gameplay/Jump/InputJump.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0e07e795727bf3546bf1f740d63d2302
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
