--- Assets/Scripts/Gameplay/Jump/InputJump/IJumpInput.cs
@@ -1,10 +1,0 @@
-﻿using System;
-
-namespace Assets.Scripts.Gameplay.InputJump.Installer
-{
-    public interface IJumpInput
-    {
-        event Action StartJump;
-        event Action CancelJump;
-    }
-}
