--- Assets/Scripts/Gameplay/Jump/InputJump/IJumpInput.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0deb6760124359843bda535bd7557f0f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
