--- Assets/Scripts/Gameplay/Jump/InputJump/InputJumpPC.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7af3a471f4fd0d94584fa3050748c123
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
